+++ packages/sdk/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.18](https://github.com/ttoss/soat/compare/v0.13.17...v0.13.18) (2026-07-03)
+
+**Note:** Version bump only for package @soat/sdk
+
+## [0.13.17](https://github.com/ttoss/soat/compare/v0.13.16...v0.13.17) (2026-07-03)
+
+**Note:** Version bump only for package @soat/sdk
+
 ## [0.13.16](https://github.com/ttoss/soat/compare/v0.13.15...v0.13.16) (2026-07-03)
 
 **Note:** Version bump only for package @soat/sdk